+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 * Switch to Alpine for smaller images
 * Switch to Quagga or Zebra deamon
+* Added BGP_IPV6 options
 
 ## 1.5.0 / 2018-05-11